codes to retrieve power info via zetsysman

+++ sycl/naxpy-sycl.cpp
@@ -47,7 +47,7 @@ static void runkernel(void)
 
 int main()
 {
-    runkernel<float>();
+    runkernel<double>();
 
     return 0;
 }
